requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9
pydantic==2.5.0
groq==0.4.2
dspy-ai==2.4.0
qdrant-client==1.7.3
sentence-transformers==2.3.1
tavily-python==0.3.3
python-dotenv==1.0.0
httpx==0.26.0
numpy==1.26.0